refactor: stream_proxy dissolution final stage - clear last file-nloc>500 flag (iter35)
Behavior-preserving completion of the stream_proxy.py god-file dissolution.
Moves the remaining large class bodies and the module-level constants into
cohesive sibling modules, leaving stream_proxy.py a thin facade (imports,
the 4 mutable-state globals + singleton plumbing, class declarations, and the
bottom re-export block):

- HlsProducer -> mixin-split via MRO into _HlsProduceMixin (stream_proxy_hls_ffmpeg)
  + _HlsSegmentMixin (stream_proxy_hls_segment); __init__ stays on HlsProducer.
- _ThreadedHTTPServer -> stream_proxy_httpserver; _ProxyStreamState -> stream_proxy_state.
- get_service_proxy_port/token/config -> stream_proxy_serviceprops.
- _get_private_hls_temp_root -> stream_proxy_ffmpeg (mutates the parent's
  _HLS_PRIVATE_TEMP_ROOT/_lock through _sp so the test patch surface is unchanged).
- All 98 module-level configuration constants -> stream_proxy_const, re-exported
  in the parent via an EXPLICIT import list (not import *, so static analysers can
  follow the re-export chain for tests' `from resources.lib.stream_proxy import _X`).

Every @patched / cross-module name stays resolvable as resources.lib.stream_proxy.<name>;
moved code reaches parent/patched names at call time via `import resources.lib.stream_proxy
as _sp` + `_sp.<name>` (preserves monkeypatch, no new import cycle). HlsProducer dir()
method set is byte-identical to HEAD (AST-verified, 36 methods); all 98 constants moved
verbatim.

Codacy (Lizard) gauge: stream_proxy.py file-nloc 524 -> well under 500; the LAST
file-nloc>500 flag is cleared. Total 24 -> 23. params>8 unchanged at 16 (not increased).
All new modules <500 NLOC.

Gated: just lint (pylint 10/10, ruff/black/vermin clean), unit suite (2138 pass;
the fallback_streams follow-up-fetch failure was a non-deterministic full-suite
concurrency load flake -- passes in isolation and on clean re-run, and is causally
unrelated to this stream_proxy-only change), integration suite (real ffmpeg, 2 pass).
Adversarially reviewed by a 3-lens panel (verbatim-behavior, name-resolution,
import-cycle/facade) -- clean, 0 missing names across 194 sibling _sp accessors,
35 @patch targets, and 46 test from-imports.
